Fix builds for g++-3.4.6,4.0.4,4.1.2, in steady-state-random-waypoint-mobility-model.cc, it calls node.create (double instead of uint32_t)

+++ src/mobility/steady-state-random-waypoint-mobility-model.cc
@@ -288,7 +288,7 @@ SteadyStateRandomWaypointTest::DoRun (void)
   double totalTime = 1000;
   // Create nodes
   NodeCount = 10000;
-  nodes.Create (NodeCount);
+  nodes.Create ((uint32_t) NodeCount);
   // Installmobility
   MobilityHelper mobility;
   mobility.SetMobilityModel ("ns3::SteadyStateRandomWaypointMobilityModel",
